After 5 multiples, no point to 32-bit bitset

+++ src/qimcifa.cpp
@@ -285,7 +285,7 @@ template <typename BigInteger> bool isMultiple(const BigInteger& p, const std::v
     return false;
 }
 
-template <typename BigInteger> boost::dynamic_bitset<uint32_t> wheel_inc(std::vector<BigInteger> primes) {
+template <typename BigInteger> boost::dynamic_bitset<uint64_t> wheel_inc(std::vector<BigInteger> primes) {
     BigInteger radius = 1U;
     for (const BigInteger& i : primes) {
         radius *= i;
@@ -301,7 +301,7 @@ template <typename BigInteger> boost::dynamic_bitset<uint32_t> wheel_inc(std::ve
         }
     }
 
-    boost::dynamic_bitset<uint32_t> output(count);
+    boost::dynamic_bitset<uint64_t> output(count);
     for (size_t i = 0U; i < count; ++i) {
         output[i] = o[i];
     }
@@ -310,8 +310,8 @@ template <typename BigInteger> boost::dynamic_bitset<uint32_t> wheel_inc(std::ve
     return output;
 }
 
-template <typename BigInteger> std::vector<boost::dynamic_bitset<uint32_t>> wheel_gen(const std::vector<BigInteger>& primes) {
-    std::vector<boost::dynamic_bitset<uint32_t>> output;
+template <typename BigInteger> std::vector<boost::dynamic_bitset<uint64_t>> wheel_gen(const std::vector<BigInteger>& primes) {
+    std::vector<boost::dynamic_bitset<uint64_t>> output;
     std::vector<BigInteger> wheelPrimes;
     for (const BigInteger p : primes) {
         wheelPrimes.push_back(p);
@@ -461,7 +461,7 @@ bool singleWordLoop(const bitCapInt& toFactor, const bitCapInt& range, const bit
 }
 #else
 template <typename bitCapInt>
-bool singleWordLoop(const bitCapInt& toFactor, std::vector<boost::dynamic_bitset<uint32_t>> inc_seqs, const std::chrono::time_point<std::chrono::high_resolution_clock>& iterClock)
+bool singleWordLoop(const bitCapInt& toFactor, std::vector<boost::dynamic_bitset<uint64_t>> inc_seqs, const std::chrono::time_point<std::chrono::high_resolution_clock>& iterClock)
 {
     for (bitCapInt batchNum = (bitCapInt)getNextBatch(); batchNum < batchBound; batchNum = (bitCapInt)getNextBatch()) {
         const bitCapInt batchStart = (bitCapInt)((batchCount - (batchNum + 1U)) * BASE_TRIALS + 2U);
@@ -481,7 +481,7 @@ bool singleWordLoop(const bitCapInt& toFactor, std::vector<boost::dynamic_bitset
 
             bool is_wheel_multiple = false;
             for (size_t i = 0; i < inc_seqs.size(); ++i) {
-                boost::dynamic_bitset<uint32_t>& wheel = inc_seqs[i];
+                boost::dynamic_bitset<uint64_t>& wheel = inc_seqs[i];
                 is_wheel_multiple = wheel[0U];
                 wheel >>= 1U;
                 if (is_wheel_multiple) {
@@ -541,7 +541,7 @@ int mainBody(const bitCapInt& toFactor, const uint64_t& tdLevel, const std::vect
 #if IS_RANDOM
     std::random_device seeder;
 #else
-    std::vector<boost::dynamic_bitset<uint32_t>> inc_seqs = wheel_gen(std::vector<bitCapInt>(trialDivisionPrimes.begin(), trialDivisionPrimes.begin() + tdLevel));
+    std::vector<boost::dynamic_bitset<uint64_t>> inc_seqs = wheel_gen(std::vector<bitCapInt>(trialDivisionPrimes.begin(), trialDivisionPrimes.begin() + tdLevel));
 #endif
 
 #if IS_PARALLEL
